+++ NEWS.md
@@ -1,6 +1,7 @@
 # santoku (development version)
 
 * Bugfix: `lbl_endpoint()` was erroring for some types of breaks.
+* Use unicode em-dash and no spaces in `lbl_dash()` and `lbl_discrete()`.
 
 # santoku 0.6.0